final financial system draft

+++ app/Http/Controllers/resultsController.php
@@ -35,7 +35,7 @@ class resultsController extends AppBaseController
     }
 
     /**
-     * Display a listing of the marks.
+     * Display a listing of The Results.
      *
      * @param Request $request
      *
@@ -45,132 +45,18 @@ class resultsController extends AppBaseController
     public function index(Request $request)
     {
 
-        $sems = sems::with('years')->get();
-        $classrooms = Classrooms::with('levels')->get();
-        $years = Years::all();
-        $levels = Levels::all();
-        $courses = Courses::all();
+        $editby = date("Y-m-d H:i:s", strtotime('-1 day', strtotime(now())));
 
-        $currentSem = DB::table('sems')->select( // Current Sem Data
-            'sems.*',
-            'years.*'
-        )
-        ->join('years', 'years.year_id', '=', 'sems.year_id')
+        $currentSem = sems::with('year')
         ->where('sems.start', '<=', today())
-        ->where('end', '>=', today())->limit(1)->get();
+        ->where('end', '>=', today())
+        ->first();
 
-        $sches = Sches::all();
-        
-        $students = DB::table('users')->select( 
-            'students.*',
-            'users.*'
-        )
-        ->join('students', 'students.studentNo', '=', 'users.schoolNo')
-        ->get();
+        $levels = levels::all();
+        $classrooms = classrooms::with('level.courses.markstypes.marks.student.user', 'level.courses.markstypes.sem.year')->get();
+        $courses = courses::all();
 
-        $semY = DB::table('sems')->select(
-            'sems.*',
-            'years.*'
-        )
-        ->join('years', 'years.year_id', '=', 'sems.year_id')
-        ->orderBy('sems.created_at', 'DESC')->get();
-
-        $marks = DB::table('marks')->select(
-            'markstypes.*',
-            'sems.*',
-            'classrooms.*',
-            'years.*',
-            'courses.*',
-            'users.*',
-            'marks.*'
-        )
-        ->join('markstypes', 'markstypes.type_id', '=', 'marks.type_id')
-        ->join('sems', 'sems.sem_id', '=', 'markstypes.sem_id')
-        ->join('classrooms', 'classrooms.classroom_id', '=', 'markstypes.classroom_id')
-        ->join('years', 'years.year_id', '=', 'sems.year_id')
-        ->join('courses', 'courses.course_id', '=', 'markstypes.course_id')
-        ->join('users', 'users.schoolNo', '=', 'marks.studentNo')
-        ->orderBy('sems.sem_id', 'DESC')->orderBy('classrooms.classroom_id', 'DESC')
-        ->orderBy('courses.course_id', 'DESC')->get();
-
-        return view('results.index', compact('sches', 'marks', 'sems', 'currentSem', 'classrooms',
-        'semY', 'levels', 'courses', 'students'));
-    }
-
-    public function currentSem(){ // Results Generation ///////////////////////////////////////////
-
-        $currentSem = DB::table('sems')->select(
-            'sems.*',
-            'years.*'
-        )
-        ->join('years', 'years.year_id', '=', 'sems.year_id')
-        ->where('sems.start', '<=', today())
-        ->where('end', '>=', today())->limit(1)->get();
-
-        return Response::json($currentSem);
-
-    }
-
-    public function alllevels(){ // Results Generation ///////////////////////////////////////////
-
-        $levels = Levels::all();
-
-        return Response::json($levels);
-
-    }
-
-    public function activeclassrooms(Request $request){ // Results Generation ///////////////////////////////////////////
-
-        $level_id = $request->get('level_id');
-
-        $classrooms = Classrooms::where('status_id', '=', 2)->where('level_id', '=', $level_id)->get();
-
-        return Response::json($classrooms);
-
-    }
-
-    public function activestudents(Request $request){ // Results Generation ///////////////////////////////////////////
-
-        $classroom_id = $request->get('classroom_id');
-
-        $students = DB::table('students')->select(
-            'users.*',
-            'students.*'
-        )
-        ->join('users', 'users.schoolNo', '=', 'students.studentNo')
-        ->where('students.status', '=', $classroom_id)->where('users.classroom_id', '=', 2)->get();
-
-        return Response::json($students);
-
-    }
-
-    public function activecourses(Request $request){ // Results Generation ///////////////////////////////////////////
-
-        $level_id = $request->get('level_id');
-
-        $courses = courses::where('status_id', '=', 2)->where('level_id', '=', $level_id)->get();
-
-        return Response::json($courses);
-
-    }
-
-    public function marksntypes(Request $request){ // Results Generation ///////////////////////////////////////////
-
-        $studentNo = $request->get('studentNo');
-        $course_id = $request->get('course_id');
-        $sem_id = $request->get('sem_id');
-
-        $marks =  DB::table('marks')->select(
-            'markstypes.*',
-            'marks.*'
-        )
-        ->join('markstypes', 'markstypes.type_id', '=', 'marks.type_id')
-        ->where('marks.studentNo', '=', $studentNo)
-        ->where('markstypes.course_id', '=', $course_id)
-        ->where('markstypes.sem_id', '=', $sem_id)->get();
-
-        return Response::json($marks);
-
+        return view('results.index', compact('editby', 'currentSem', 'classrooms', 'levels', 'courses'));
     }
 
     /**
@@ -183,11 +69,20 @@ class resultsController extends AppBaseController
 
     public function store(CreatemarksRequest $request)
     {
-        $input = $request->all();
+        $this->authorize('create', marks::class);
+
+        $classroom = $request['classroom'];
+
+        $currentSem = sems::
+        where('start', '<=', today())
+        ->where('end', '>=', today())
+        ->first();
+
+        $marks = markstypes::where('sem_id', '=', $currentSem['id']);
 
         $marks = $this->marksRepository->create($input);
 
-        Flash::success('The results were saved successfully<br><br>تم حفظ النتائج بنجاح');
+        Flash::success('The results of '.$classroom.' were created successfully<br><br> تم إصدار النتائج للصف الدراسي'.$classroom.' بنجاح');
 
         return redirect(route('results.index'));
     }
@@ -203,19 +98,21 @@ class resultsController extends AppBaseController
 
     public function update(Request $request) // Updating with Modal
      {
-         $mark = marks::findOrFail($request->mark_id);
+        $this->authorize('update', marks::class);
 
-         if (empty($mark)) {
-             Flash::error('The mark was not found<br><br>بيانات العلامة المطلوبة غير موجودة');
+        $mark = marks::findOrFail($request->mark_id);
 
-             return redirect(route('results.index'));
-         }
-        
-         $mark->update($request->all());
+        if (empty($mark)) {
+            Flash::error('The Result was not found<br><br>بيانات النتيجة المطلوبة غير موجودة');
 
-         Flash::success('The mark was updated successfully<br><br>تم تحديث بيانات العلامة بنجاح');
+            return redirect(route('results.index'));
+        }
+    
+        $mark->update($request->all());
 
-         return redirect(route('results.index'));
+        Flash::success('The Result was updated successfully<br><br>تم تحديث بيانات النتيجة بنجاح');
+
+        return redirect(route('results.index'));
      }
 
     /**
@@ -230,19 +127,21 @@ class resultsController extends AppBaseController
 
     public function destroy(Request $request)
     {
+        $this->authorize('delete', marks::class);
+
         $id = $request['id'];
         
         $marks = $this->marksRepository->find($id);
 
         if (empty($marks)) {
-            Flash::error('The mark was not found<br><br>بيانات العلامة المطلوبة غير موجودة');
+            Flash::error('The Result was not found<br><br>بيانات النتيجة المطلوبة غير موجودة');
 
             return redirect(route('results.index'));
         }
 
         $this->marksRepository->delete($id);
 
-        Flash::success('The mark was deleted successfully<br><br>تم حذف بيانات العلامة بنجاح');
+        Flash::success('The Result was deleted successfully<br><br>تم حذف بيانات النتيجة بنجاح');
 
         return redirect(route('results.index'));
     }
